# docs/openapi.yaml
...
openapi: 3.0.1
info:
  title: SEGS API
  description: 'Secure Email Gateway Suite HTTP API. Authenticate with a stateful JWT (``Authorization:
    Bearer``) or the ``seg_session`` cookie. SCIM 2.0 is at ``/scim/v2``. Interactive ``/docs``
    is not served; the contract is ``docs/openapi.yaml`` and is what API Gateway imports.'
  version: 1.0.0
  x-generated-by: backend.api.openapi
servers:
- url: /
  description: Same origin as the SOC console (CloudFront)
paths:
...
  /api/users:
    get:
      summary: List Users
      operationId: list_users_api_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - bearerAuth: []
      - cookieAuth: []
    post:
      summary: Create User
      operationId: create_user_api_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/users/{user_id}:
    delete:
      summary: Delete User
      operationId: delete_user_api_users__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
...
  /scim/v2/ServiceProviderConfig:
    get:
      tags:
      - scim
      summary: Service Provider Config
      operationId: service_provider_config_scim_v2_ServiceProviderConfig_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - scimBearer: []
  /scim/v2/ResourceTypes:
    get:
      tags:
      - scim
      summary: Resource Types
      operationId: resource_types_scim_v2_ResourceTypes_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - scimBearer: []
  /scim/v2/Schemas:
    get:
      tags:
      - scim
      summary: Schemas
      operationId: schemas_scim_v2_Schemas_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - scimBearer: []
  /scim/v2/Users:
    get:
      tags:
      - scim
      summary: List Users
      operationId: list_users_scim_v2_Users_get
      parameters:
      - name: filter
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Filter
      - name: startIndex
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Startindex
      - name: count
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Count
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - scimBearer: []
    post:
      tags:
      - scim
      summary: Create User
      operationId: create_user_scim_v2_Users_post
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - scimBearer: []
  /scim/v2/Users/{user_id}:
    get:
      tags:
      - scim
      summary: Get User
      operationId: get_user_scim_v2_Users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - scimBearer: []
    put:
      tags:
      - scim
      summary: Replace User
      operationId: replace_user_scim_v2_Users__user_id__put
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - scimBearer: []
    patch:
      tags:
      - scim
      summary: Patch User
      operationId: patch_user_scim_v2_Users__user_id__patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - scimBearer: []
    delete:
      tags:
      - scim
      summary: Delete User
      operationId: delete_user_scim_v2_Users__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - scimBearer: []
  /scim/v2/Groups:
    get:
      tags:
      - scim
      summary: List Groups
      operationId: list_groups_scim_v2_Groups_get
      parameters:
      - name: filter
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Filter
      - name: startIndex
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Startindex
      - name: count
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Count
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - scimBearer: []
  /scim/v2/Groups/{group_id}:
    get:
      tags:
      - scim
      summary: Get Group
      operationId: get_group_scim_v2_Groups__group_id__get
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - scimBearer: []
    patch:
      tags:
      - scim
      summary: Patch Group
      operationId: patch_group_scim_v2_Groups__group_id__patch
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - scimBearer: []
...
components:
  schemas:
...
    CreateUserRequest:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 1
          pattern: ^[a-zA-Z0-9._@\-]+$
          title: Username
        password:
          type: string
          maxLength: 256
          minLength: 8
          title: Password
        role:
          type: string
          title: Role
      type: object
      required:
      - username
      - password
      - role
      title: CreateUserRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateful JWT (RFC 6750). Same token as the seg_session cookie; jti is stored
        server-side.
    cookieAuth:
      type: apiKey
      in: cookie
      name: seg_session
      description: Session cookie set at login. Same credential as Bearer.
    scimBearer:
      type: http
      scheme: bearer
      description: SCIM 2.0 bearer (SEG_SCIM_BEARER_TOKEN) or an admin JWT.